move websocket code to a seperate module
